codeintel: Pull background tasks into dependencies service  (#42535)
