+++ cmd/worker/internal/codeintel/crates_syncer.go
@@ -6,7 +6,9 @@ import (
 	"github.com/sourcegraph/log"
 
 	"github.com/sourcegraph/sourcegraph/cmd/worker/job"
+	"github.com/sourcegraph/sourcegraph/cmd/worker/shared/init/codeintel"
 	workerdb "github.com/sourcegraph/sourcegraph/cmd/worker/shared/init/db"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/background/cratesyncer"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/env"
@@ -27,7 +29,12 @@ func (j *cratesSyncerJob) Routines(startupCtx context.Context, logger log.Logger
 	}
 	db := database.NewDB(logger, rawDB)
 
-	return []goroutine.BackgroundRoutine{
-		cratesyncer.NewCratesSyncer(db),
-	}, nil
+	gitserverClient, err := codeintel.InitGitserverClient()
+	if err != nil {
+		return nil, err
+	}
+
+	depsSvc := dependencies.GetService(db, gitserverClient)
+
+	return cratesyncer.NewSyncers(depsSvc), nil
 }
