+++ internal/codeintel/dependencies/background/cratesyncer/iface.go
@@ -1,0 +1,7 @@
+package cratesyncer
+
+import "github.com/sourcegraph/sourcegraph/internal/goroutine"
+
+type DependenciesService interface {
+	NewCrateSyncer() goroutine.BackgroundRoutine
+}
